Added ids to each page
Mostly spent the majority of the time figuring out how the HTML page ids worked. Learned that using the hash would allow you to go to a tag id whenever specified, and the id is written as equal to a value in an attribute. Furthermore, I also learned how to link to an id on another page through trial and error.